// elohim-app/src/app/shefa/interfaces/requests-and-offers.interface.ts
export interface IRequestsAndOffers {
// ...
  ): Promise<{
    request: ServiceRequest;
    updatedEvent: EconomicEvent;
  }>;
// ...
    filters?: {
      status?: 'active' | 'archived' | 'deleted';
      fromDate?: string;
      toDate?: string;
    }
// ...
}